Managed policies for dynamo
